# PBE/api-produtos-multicamadas/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Produtos Multicamadas
  version: 1.0.0
  description: CRUD completo de produtos com Node.js, Express, Sequelize e PostgreSQL.
servers:
  - url: http://localhost:3000
paths:
...
  /api/produtos:
    get:
      summary: Lista produtos com filtros e paginação
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: sku
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: minPrice
          schema: { type: number, minimum: 0 }
        - in: query
          name: maxPrice
          schema: { type: number, minimum: 0 }
          description: Deve ser maior ou igual a minPrice quando ambos forem informados.
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [name, sku, category, price, stock, active, createdAt, updatedAt]
            default: createdAt
        - in: query
          name: order
          schema:
            type: string
            enum: [ASC, DESC, asc, desc]
            default: DESC
      responses:
        '200':
          description: Lista retornada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Filtros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Cria um produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: SKU duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ProductCreate:
      type: object
      required: [sku, name, category, price]
      properties:
        sku: { type: string, minLength: 2, maxLength: 50, example: MON-001 }
        name: { type: string, minLength: 2, maxLength: 150, example: Monitor 24 polegadas }
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Monitor IPS Full HD
        category:
          type: string
          minLength: 2
          maxLength: 100
          example: Monitores
        price:
          type: number
          minimum: 0
          maximum: 9999999999.99
          example: 899.90
        stock:
          type: integer
          minimum: 0
          maximum: 2147483647
          default: 0
          example: 15
        active: { type: boolean, default: true }
...
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductCreate'
        - type: object
          required: [id, stock, active, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    ProductResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          $ref: '#/components/schemas/Product'
    Pagination:
      type: object
      required: [page, limit, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    ProductListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }